调整代码结构，方便后续刷新页面
