+++ run.py
@@ -5,6 +5,8 @@ sys.path.append("../")
 from PyQt5 import QtWidgets
 
 from main.CaseModel import Ui_Form_Main
+from main.SublevelHelp import Ui_Form_Help
+from util.logger import logger
 
 
 class Run(QtWidgets.QWidget):
@@ -12,6 +14,16 @@ class Run(QtWidgets.QWidget):
         super(Run, self).__init__()
         self.ui = Ui_Form_Main()
         self.ui.setupUi(self)
+        self.ui.pushButton_sublevel_help.clicked.connect(self.sublevel_help)
+
+    # 打开次层元素帮助方法
+    def sublevel_help(self):
+        try:
+            self.help_widget = QtWidgets.QWidget()
+            self.window_help = Ui_Form_Help()
+            self.window_help.setupUi(self.help_widget)
+        except:
+            logger.exception("发现错误:")
 
 
 if __name__ == "__main__":
